Update Long uid

+++ core-data/src/main/java/com/taekwondo/coredata/network/dao/EventDao.kt
@@ -19,5 +19,5 @@ interface EventDao {
     fun getAllEvents(): List<EventEntity>
 
     @Query("SELECT * FROM event WHERE eventId = :uid")
-    fun getEvent(uid: Int): EventEntity?
+    fun getEvent(uid: Long): EventEntity?
 }
