+++ feature-event/src/main/java/com/taekwondo/featureevent/presentation/model/FighterModel.kt
@@ -1,8 +1,8 @@
 package com.taekwondo.featureevent.presentation.model
 
 data class FighterModel(
-    val uid: Int,
-    val name:String,
+    val uid: Long,
+    val name: String,
     val photo: String?,
     val isPicked: Boolean,
 )
